Add IconsInfo and Projects Widgets to MainSection
Created new IconsInfo and Projects widgets to display information and projects
in the main section of the application. The IconsInfo widget shows statistics
like clients, projects, countries, and stars using icons and text. The Projects
widget displays a grid of project items with images, titles, descriptions, and
a "More Info" button. Updated the MainSection to include these new widgets.